requirements.txt
python-ffmpeg
yt-dlp
openai-whisper
